Roteiros/Roteiro 2/app_recon/requirements.txt
whois==0.7.3
dnspython==2.3.0
sublist3r==1.0.0
wafw00f==2.1.0
inquirer==2.7.0
pytest==7.2.0